.NET: Refine streamed annotation correlation
Copilot-Session: 0c67da26-d12e-47ab-aac4-53e251ef6edf
